refactor: split pictures file
